Update gff2sql.nf
Adding a conditional to check with digIS file is empty

+++ modules/generic/gff2sql.nf
@@ -15,13 +15,25 @@ process create_sql {
 
   script:
   """
-  # concatenate files
-  cat $gff digIS.gff | bedtools sort > input.gff
-  cat $genes_nt digIS.fa  > input.fa
-  cat $genes_aa digIS.faa > input.faa
+  if [ -s digIS.fa ] ;
+  then
 
-  # Create SQL db
-  gff2sql.R -i input.gff -o ${prefix}.sqlite -n input.fa -a input.faa -f $genome &> gff2sql.log;
+    # concatenate files
+    cat $gff digIS.gff | bedtools sort > input.gff
+    cat $genes_nt digIS.fa  > input.fa
+    cat $genes_aa digIS.faa > input.faa
+
+    # Create SQL db
+    gff2sql.R -i input.gff -o ${prefix}.sqlite -n input.fa -a input.faa -f $genome &> gff2sql.log;
+
+  else
+
+    # Create SQL db
+    gff2sql.R -i $gff -o ${prefix}.sqlite -n $genes_nt -a $genes_aa -f $genome &> gff2sql.log;
+
+  fi
+
+  # Save results with better name
   mv /work/${prefix}.sqlite . ;
 
   # Save parser
